Error-logging: use WC Logger when possible

+++ includes/class-wcmp-export.php
@@ -1428,15 +1428,21 @@ class WooCommerce_MyParcel_Export {
 
 	public function log( $message ) {
 		if (isset(WooCommerce_MyParcel()->general_settings['error_logging'])) {
-			// log file in upload folder - wp-content/uploads
-			$upload_dir = wp_upload_dir();
-			$upload_base = trailingslashit( $upload_dir['basedir'] );
-			$log_file = $upload_base.'myparcel_log.txt';
+			if( class_exists('WC_Logger') ) {
+				$wc_logger = function_exists('wc_get_logger') ? wc_get_logger() : new WC_Logger();
+				$wc_logger->add('wc-myparcel', $message );
+			} else {
+				// Old WC versions didn't have a logger
+				// log file in upload folder - wp-content/uploads
+				$upload_dir = wp_upload_dir();
+				$upload_base = trailingslashit( $upload_dir['basedir'] );
+				$log_file = $upload_base.'myparcel_log.txt';
 
-			$current_date_time = date("Y-m-d H:i:s");
-			$message = $current_date_time .' ' .$message ."\n";
+				$current_date_time = date("Y-m-d H:i:s");
+				$message = $current_date_time .' ' .$message ."\n";
 
-			file_put_contents($log_file, $message, FILE_APPEND);
+				file_put_contents($log_file, $message, FILE_APPEND);
+			}
 		}
 	}
 }
